feat: compare optimal filter on real-world data (#21)
Initialize real-world experiments based on the denoising task.
Add utilization of sea temperature data and sample filtering.
Update the optimal filter generation process with the correct
output. Add a maximum limit to node and time instance count
through initialization of the graph and joint time-vertex signal.
Remove the artifacts of the real-world main, which were related
to the full version of the real-world dataset. Add the GPU
version of the optimal filter generation from the
`real-world-gpu` branch.

Implement optimal filtering experiment by sampling a small
subset from the real-world datasets. The auto- and
cross-correlation are assumed to be known for the joint
time-vertex signal and joint time-vertex noise. The optimal
filter is obtained from these correlations and applied to the
noisy version.

Add a progress bar to track the progress of the grid search.
Update the experiment parameters to decrease the NT
multiplication but enhance the grid search.

Add optimal filtering results for several configurations
with 2-NN graphs, where the node counts are `N=4,5,8`
and `T=8,5,8`, correspondingly.

Add a result analyzer and info extractor for the
real-world optimum filtering task. Update display information
of optimal real results to be more structured.

Add median filtering comparison to the real optimal filtering
comparison. Add the ARMA version to the optimal sea surface
temperature dataset filtering task.

Add optimal filtering results on PM-2.5 dataset with on
the computationally most expensive setting of N=8, T=8.

Add remaining PM-2.5 datasets results for N=4,5 and T=8,5
correspondingly. Move the sea surface temperature results
of the optimal filtering task to their directory.

For the analysis, the function fixes the path of the dataset.
The optimal filtering task comparison was handled
under the `real-world` directory residing under the root
directory. This directory is moved under the optimal
filtering task.

The synthetic optimal filtering task files residing under the
moved directory are also moved to the `synthetic` directory.

Refs: #8
Refs: #14
Resolves: #19
Resolves: #20